Add point-to-road snapping
In general, when a user specifies a point on a map either by clicking
or by dragging a marker, it will be near the road, but not exactly on it.
Often, it may land in courtyard, a pedestrian zone, etc, and then the
routing engine may not be able to display a valid route.
The snapping endpoint accepts the user-specified point and returns the
closest point on the closest primary, secondary or tertriary road. For
the exact definitions of these, see:
https://wiki.openstreetmap.org/wiki/Key:highway

endpoint: /snap
method: GET
query parameter: point <lat,lon>
returns: <lon,lat>
e.g. /snap/?point=55.752908,37.613011

Search radius is hardcoded as 1km. Within the Moscow Region, this should
sufficient, otherwise the user may have dropped the point erroneously.