Allow intermediate waypoints in _/directions_
Previously, there were _start_ & _end_ params, now there's a single
_positions_ param that accepts a semicolon-delimited list of positions:
    _positions=x0,y0;x1,y1;x2,y2;..._

The positions __must__ be ordered.
Whether a single route or multiple alternatives are returned, depends on
the number of positions in the request. If there are two, a (for now)
hardcoded default of three routes is returned. If there are more positions
supplied, only one route is returned.

Tests are added as well.