Add PostGIS as service in CI tests
Since snapping is now part of the directions endpoint, a DB with roads
is required to test it.
For now, our "custom" PostGIS with preloaded roads is used. But that means that the roads are for
Moscow & Moscow Region but ORS graph are the default Heidelberg ones. In the future, a better solution should be found.