feat(pickup): return passenger route distance
Previously, only the straight-line distance was returned. To aid the front-end logic,
'distance' was renamed to 'radius' in the response, and a new 'distance' key was added
whose value is the along-the-graph pedestrian route distance.

Another imporant change: planar calculations are now in UTM 37U, which is the UTM zone for Moscow,
since it turned out that the distances in 3857 were heavily distorted.