Add Immitate
For trsting purposes, it's useful to be able to immitate a realistic driver's route,
with the slight deviations brought about by the GPS inaccuracy.
The endpoint accepts an existing route and randomly shift each of its vertices by up to 20m.
It also allows supplying a set of points to deviate to. They will be used to replace
the point on the exisiting route that are the closest to them.