Change Pickup Point model
It turned out more convenient to identify associate a pick-up point with
it's route, not the trip or the user. Hence, the trip_id & user_id fields
are now gone and replaced by a single route_id which is also a backref to
the route table (PickupPoint.route returns the user's route).

This change entailed a change in the corresponding POST endpoint:
a route_id is now accepted instead of trip_id & user_id.